src/main/java/net/engineeringdigest/journalApp/service/UserService.java
package net.engineeringdigest.journalApp.service;

import lombok.extern.slf4j.Slf4j;
import net.engineeringdigest.journalApp.entity.User;
import net.engineeringdigest.journalApp.repository.UserRepository;
import org.bson.types.ObjectId;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.stereotype.Component;

import java.util.Arrays;
import java.util.List;
import java.util.Optional;

@Slf4j
@Component
public class UserService {

    private static PasswordEncoder passwordEncoder = new BCryptPasswordEncoder();

    // @Slf4j helps us to remove this instance we can directly use log without writing this line
//    private static Logger log = LoggerFactory.getLogger(UserService.class);
    @Autowired
    private UserRepository userRepository;

    public void saveUser(User user) {

        userRepository.save(user);


    }

    public boolean saveNewUser(User user) {
        try {
            user.setPassword(passwordEncoder.encode(user.getPassword()));
            user.setRoles(Arrays.asList("USER"));
            userRepository.save(user);
            return true;
        } catch (Exception e) {
            // default loggers
            log.info("info log");
            log.warn("warn occurred for {}", user.getUserName(), e);
            log.error("error occurred for {}", user.getUserName(), e);
            // not default loggers
            log.debug("debug log");
            log.trace("trace logger");
            return false;
        }
    }

    public void createAdmin(User user) {
        user.setPassword(passwordEncoder.encode(user.getPassword()));
        user.setRoles(Arrays.asList("USER", "ADMIN"));
        userRepository.save(user);
    }

    public List<User> getAllUsers() {
        return userRepository.findAll();
    }

    public User findByUserName(String userName) {
        return userRepository.findByUserName(userName);
    }


    public void deleteById(ObjectId myId) {
        userRepository.deleteById(myId);
    }
}
